requirement.txt
google-genai
streamlit
yfinance
pandas
numpy
ta